Move target tracking PID to Drive
