+++ robot/src/main/java/us/ilite/robot/commands/TargetLock.java
@@ -1,7 +1,5 @@
 package us.ilite.robot.commands;
 
-import com.ctre.phoenix.motorcontrol.ControlMode;
-import com.ctre.phoenix.motorcontrol.NeutralMode;
 import com.flybotix.hfr.codex.Codex;
 
 import us.ilite.common.lib.util.CheesyDriveHelper;
@@ -21,20 +19,14 @@ import us.ilite.robot.modules.targetData.ITargetDataProvider;
 
 public class TargetLock implements ICommand {
 
-    private static final double kMIN_POWER = -1.0;
-    private static final double kMAX_POWER = 1.0;
-    private static final double kMIN_INPUT = -27;
-    private static final double kMAX_INPUT = 27;
     private static final double kTURN_POWER = 0.2;
     private static final int kAlignCount = 10;
-    private static final double kFrictionFeedforward = 0.44 / 12;
     private static final double kTargetAreaScalar = 1.0;
 
     private Drive mDrive;
     private ITargetDataProvider mCamera;
     // Different throttle providers give us some control over behavior in autonomous
     private IThrottleProvider mTargetSearchThrottleProvider, mTargetLockThrottleProvider;
-    private PIDController mPID;
     private ETrackingType mTrackingType;
 
     private double mAllowableError, mPreviousTime, mOutput = 0.0;
@@ -42,8 +34,6 @@ public class TargetLock implements ICommand {
     private boolean mEndOnAlignment = true;
     private int mAlignedCount = 0;
     private boolean mHasAcquiredTarget = false;
-
-    private CheesyDriveHelper mCheesyDriveHelper = new CheesyDriveHelper(SystemSettings.kCheesyDriveGains);
 
     public TargetLock(Drive pDrive, double pAllowableError, ETrackingType pTrackingType, ITargetDataProvider pCamera, IThrottleProvider pThrottleProvider) {
         this(pDrive, pAllowableError, pTrackingType, pCamera, pThrottleProvider, true);
@@ -62,43 +52,33 @@ public class TargetLock implements ICommand {
     @Override
     public void init(double pNow) {
         System.out.println("++++++++++++++++++++++++++TARGET LOCKING++++++++++++++++++++++++++++++++++++\n\n\n\n");
+
         mHasAcquiredTarget = false;
         mAlignedCount = 0;
-        mPID = new PIDController(SystemSettings.kTargetAngleLockGains, kMIN_INPUT, kMAX_INPUT, SystemSettings.kControlLoopPeriod);
-        mPID.setOutputRange(kMIN_POWER, kMAX_POWER);
-        mPID.setSetpoint(0);
-        mPID.reset();
-        SmartDashboard.putBoolean("Initializing Command", true);
+
+        mDrive.setTargetTracking();
+        mDrive.setTargetTrackingThrottle(0);
 
         this.mPreviousTime = pNow;
     }
 
     @Override
     public boolean update(double pNow) {
-        SmartDashboard.putBoolean("Initializing Command", false);
         Codex<Double, ETargetingData> currentData = mCamera.getTargetingData();
-        System.out.println("LOCKING " + currentData.get(ETargetingData.tx));
 
-        if(mPID != null && currentData != null && currentData.isSet(ETargetingData.tv) && currentData.get(ETargetingData.tx) != null) {
+        if(currentData != null && currentData.isSet(ETargetingData.tv) && currentData.get(ETargetingData.tx) != null) {
             mHasAcquiredTarget = true;
 
-            //if there is a target in the limelight's fov, lock onto target using feedback loop
-            mOutput = mPID.calculate(-1.0 * currentData.get(ETargetingData.tx), pNow - mPreviousTime);
-            mOutput = mOutput + (Math.signum(mOutput) * kFrictionFeedforward);
+            mDrive.setTargetTrackingThrottle(mTargetLockThrottleProvider.getThrottle() * SystemSettings.kSnailModePercentThrottleReduction);
 
-            double throttle = mTargetLockThrottleProvider.getThrottle() * SystemSettings.kSnailModePercentThrottleReduction;
-
-            mDrive.setDriveMessage(getClampedTurnDrive(throttle, mOutput, currentData));
-
-            SmartDashboard.putNumber("PID Turn Output", mOutput);
             mAlignedCount++;
             if(mEndOnAlignment && Math.abs(currentData.get(ETargetingData.tx)) < mAllowableError && mAlignedCount > kAlignCount) {
-                //if x offset from crosshair is within acceptable error, command TargetLock is completed
                 System.out.println("FINISHED");
-                mDrive.setDriveMessage(DriveMessage.kNeutral);
+                // Zero drive outputs in shutdown()
                 return true;
             }
-            // If we've already seen the target and lose tracking, exit.
+
+        // If we've already seen the target and lose tracking, exit.
         } else if(mHasAcquiredTarget && !currentData.isSet(ETargetingData.tv)) {
             mDrive.setDriveMessage(DriveMessage.kNeutral);
             return true;
@@ -124,62 +104,8 @@ public class TargetLock implements ICommand {
 
     @Override
     public void shutdown(double pNow) {
-
-    }
-
-    /*
-    Uses CheesyDrive to generate steering and throttle commands.
-    This has the advantage of having derivative control (through "negative inertia")
-    Clamps the maximum curvature of the drivetrain (wheel sensitivity), allowing the PID to be overagressive and compensate better for large lateral offsets
-    Scales PID output by throttle, giving us better performance at low speeds
-     */
-    private DriveMessage getCheesyDrive(double throttle, double turn, Codex<Double, ETargetingData> targetData) {
-        boolean isQuickTurn = Math.abs(throttle) < Util.kEpsilon;
-
-        DriveSignal cheesyOutput = mCheesyDriveHelper.cheesyDrive(throttle, turn, false);
-        return new DriveMessage(cheesyOutput.getLeft(), cheesyOutput.getRight(), ECommonControlMode.PERCENT_OUTPUT).setNeutralMode(ECommonNeutralMode.BRAKE);
-    }
-
-    /*
-    Implements the same clamping function as CheesyDrive.
-    If throttle + turn saturates the output, the turn power being lost is applied to the other side of the drivetrain.
-    This should be better when tracking targets at high speeds.
-     */
-    private DriveMessage getClampedTurnDrive(double throttle, double turn, Codex<Double, ETargetingData> targetData) {
-
-        double leftPwm = throttle + turn;
-        double rightPwm = throttle - turn;
-
-        if (leftPwm > 1.0) {
-            rightPwm -=  (leftPwm - 1.0);
-            leftPwm = 1.0;
-        } else if (rightPwm > 1.0) {
-            leftPwm -=  (rightPwm - 1.0);
-            rightPwm = 1.0;
-        } else if (leftPwm < -1.0) {
-            rightPwm +=  (-1.0 - leftPwm);
-            leftPwm = -1.0;
-        } else if (rightPwm < -1.0) {
-            leftPwm +=  (-1.0 - rightPwm);
-            rightPwm = -1.0;
-        }
-
-        return new DriveMessage(leftPwm, rightPwm, ECommonControlMode.PERCENT_OUTPUT).setNeutralMode(ECommonNeutralMode.BRAKE);
-    }
-
-    /*
-    Implements the same scaling function as CheesyDrive, where turn is scaled by throttle.
-    This *should* give us better performance at low speeds + the benefits of "clamped turn" drive.
-     */
-    private DriveMessage getCurvatureDrive(double throttle, double turn, Codex<Double, ETargetingData> targetData) {
-        double adjustedTurn = Math.abs(throttle) * turn * SystemSettings.kTurnSensitivity;
-
-        return DriveMessage.fromThrottleAndTurn(throttle, adjustedTurn).setNeutralMode(ECommonNeutralMode.BRAKE);
-    }
-
-    private DriveMessage getArcadeDrive(double throttle, double turn, Codex<Double, ETargetingData> targetData) {
-//        mOutput *= targetData.get(ETargetingData.ta) * kTargetAreaScalar;
-        return DriveMessage.fromThrottleAndTurn(throttle, turn).setNeutralMode(ECommonNeutralMode.BRAKE);
+        mDrive.setNormal();
+        mDrive.setDriveMessage(DriveMessage.kNeutral);
     }
 
     public TargetLock setTargetLockThrottleProvider(IThrottleProvider pThrottleProvider) {
